Separate core and std modules
The core module only contains the tiny kernel pull parser and common
types. This kernel is intendend to be used to implement higher-level XML
libraries including:

 - a DOM;
 - XML document manipulation utilities;
 - XML validation;
 - XPath;
 - and so on.

Default implementation for all these features is provided by the std
module where the DOM is represented as a tree and procesors are used to
perform various validations.

The pull parser is also more strict on syntax rules.
It won’t accept new tags after the root has been closed,
according to the `document` syntax rule of the specification.